Moved functional batch tests to dedicated directory
